Update PHPUnit from 4 to 8
PHPUnit 9 can’t be used yet because it requires PHP 7.3 or later, and we
still support PHP 7.2.

Remove wmde/hamcrest-html-matchers, since it doesn’t seem to have been
used at all.

Stop using the prophecy framework; PHPUnit prints a warning that it will
be removed in version 10 (we could bring it in as an explicit trait from
a separate package, but I don’t see the need to).

Use assertion methods on the TestCase class instead of top-level
functions; the top-level functions still work, but result in warnings
that the test has no assertions, for some reason. While we’re at it,
also switch the assertions from “equal” to “same”.