Parse method calls more leniently (#26)
Allow the expression in the single argument to be anything at all, as
long as there’s a matching closing parenthesis at the end. Because we
don’t support binary operators or multiple arguments, we don’t need to
worry about counting balanced parentheses or anything like that: if the
close-paren at the very end of the string doesn’t belong to the
open-paren at the beginning of the method call, then there must be a
syntax error (or unsupported syntax) somewhere.