Support single-file component inputs
If the input contains a single root <template> and the other root
elements are all <script> or <style>, then pick the <template> contents
as the actual root element and only render its contents. This allows us
to use SFC *.vue files as the input.

It’s worth noting that we can’t use getBodyElement() directly, as I had
originally hoped – if the SFC starts with a <script> and then has a
<template> afterwards, rather than starting with the <template>, then
the PHP HTML parser will actually place the whole contents inside a
synthetic <head> rather than a synthetic <body>. (Such an SFC structure
is disallowed by eslint-config-wikimedia via vue/component-tags-order,
but I don’t think this library should rely on that.)

Bug: T395802