Extract HtmlParser class
Component’s responsibilities now start at the DOMElement; HtmlParser
contains the code to parse the HTML and extract the root node from the
DOMDocument. (I want to extend the latter logic in a follow-up commit,
and for this it’s nicer if it’s a separate class, IMHO.)

In principle, Component::render() should probably operate on a deep
clone of the root node; then a Component could be reused without having
to parse it several times, which could come in handy later. However, at
the moment, the Component is always freshly created, so the extra clone
seems wasteful; so for now just document that render() shouldn’t be
called repeatedly. The only “public” entry point to this library so far
is Templating, which obeys this rule.

Also add covers tags to TemplatingTest, since it’s kind of an
integration test for much of the library (including JsParsing, but I
can’t be bothered to add all those classes to the covers as well).

Bug: T395802