Extract HtmlParser::getBodyElement()
So we can also use it in Component::appendHTML() (v-html="" handling).
Part of this will also be useful when getRootNode() gains support for
SFC inputs later (which is why appendHTML() doesn’t use getRootNode()
directly).

In theory, we could skip over a <head> element in the <html>, but as I
don’t think it makes sense to have a <head> in any Vue component /
template, let’s just throw an error in that case and include a test for
that. (I don’t think we can test the “expected <html>” condition, that
part is basically dead code – the parser should always synthesize <html>
and <body> elements surrounding “normal” markup if it’s not already
present in the source.)