SKETCH: Introduce App class
The App class lets us reuse the same HtmlParser and JsExpressionParser
across multiple components (note that JsExpressionParser returns an
unevaluated expression, so it’s fine to cache its results even if
they’re going to be evaluated with different data later), and also keeps
a library of components. The components are loaded and compiled
on-demand the first time they’re used, and reused afterwards; for this,
Component now clones its root node before operating on it.

For Component to work correctly, it’s important that the cloned node
still has a parent node (otherwise it’s considered to have been removed
from the DOM, and most of the processing for it is skipped); use a
cloneOwner node for that purpose. (For now, it’s just the document’s
document element, i.e. the <html>, but we could make it a separate node
later if we prefer. It just has to be outside the root node.)

(If you’re wondering, my motivation for making the components input
either string or callable is that it probably makes sense not to load
all the component files into strings upfront, but to only do that if the
component is needed.)

TODO: While all the existing tests pass unmodified, which is nice, we
should also have some new tests for App – for instance, that the cloning
actually does its job (i.e. evaluating the same component in the same
app with different data yields the expected different results).