Support sub-components
This is a bit odd in that Component doesn’t even know which components
are registered in the app; sub-component handling is triggered iff the
tag name contains a hyphen. I think this is reasonable, but we might
still end up changing it.

Note that the sub-component must be a descendent of the template root
note; a template cannot just consist of a call to a sub-component. (If
it does, Component::render() crashes when trying to remove the $rootNode
from the cloneOwner because it’s already been replaced in
handleComponent().)