Support @ shorthand for v-on: directive
This is used in some WikibaseLexeme templates (albeit never in the
server-rendered path as far as I can tell), so we should support it.
There are two related problems: old versions of libxml2 complain about
@-prefixed attribute names, and our code only checked for v-on:. Both
are fairly simple to resolve.

The libxml2 error doesn’t happen on my system, so I assume it was fixed
in some recent-ish version; libxml2 2.14.0 [1], which boasts a tokenizer
“conform[ing] fully to HTML5” and “several non-standard syntax warnings
[…] removed”, seems like the most likely candidate, but I haven’t tested
the behavior with different versions. But in any event, libxml2 is
clearly able to parse the markup and produce a DOM that we can work with
(as evidenced by the fact that the WikibaseLexeme template have worked
for years before we started raising the libxml2 errors as exceptions in
ba3f8e0bef / #32), so we can just ignore the errors.

[1]: https://download.gnome.org/sources/libxml2/2.14/libxml2-2.14.0.news