Handle kebab-case attributes in component calls
libxml2 throws away the capitalization of attribute names (we always see
them all-lowercase), so component props must be specified in kebab-case
(e.g. :property-url= rather than :propertyUrl=); for this to work,
php-vuejs-templating needs to map the prop name back to camelCase
somewhere. Do this in handleComponent(). (Alternatively, it could
probably happen in App::renderComponentToDOM() or Component::render(),
but that would suggest that PHP entry points can provide props in either
format, and that feels bad to me.)