Handle member expressions in v-for statements (#40)
We currently support iterating over arrays in Vue statements, like
`<a v-for="item in list">`, but this support is currently limited
to data at the top-level of our template data object.

Add support for member expressions in v-for statements so that we
can handle statements like `<a v-for="item in list.data">`. While
we do that, make sure that ':key' attributes in 'v-for' directives
are not written out to the output.

Bug: T398011