Handle complex expressions in for loops and member expressions (#46)
In order to render structured data in Vue templates, we need to be
able to destructure more complex Javascript expressions both in the
target of 'v-for' statements and in Javascript Member Expressions
that use 'Computed' values.

Here we add support for data structures that use hyphens in key names
`<div v-for="item in list['data-values']">`, or numeric key names
`<div v-for="item in list[1]">`, or in the case where the lookup
key is itself an identifier `{{ data[variable] }}`.

Bug: T396098