Don’t clone node in handleRawHtml()
I don’t know why the code cloned the node before modifying it – there’s
no explanation of this in #2, and the rest of the code is happy to
modify nodes directly. However, this causes attribute binding on nodes
with v-html to have no effect – after handleRawHtml(), the original node
is skipped for further processing (isRemovedFromTheDom() returns true),
and even if it wasn’t, handling attributes on it wouldn’t have any
effect on the node that’s actually in the DOM now. Given that removing
the cloning doesn’t seem to break any tests, let’s just remove it and
add a test for attribute binding in conjunction with v-html.

Bug: T396633
Bug: T396634